refactor: move qr-code to TransactionView+Model
