+++ app/src/main/java/org/feature/fox/coffee_counter/ui/profile/ProfileViewModel.kt
@@ -1,14 +1,11 @@
 package org.feature.fox.coffee_counter.ui.profile
 
-import android.graphics.Bitmap
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.google.zxing.BarcodeFormat
-import com.google.zxing.qrcode.QRCodeWriter
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.receiveAsFlow
@@ -31,13 +28,11 @@ interface IProfileViewModel : IToast {
     val isAdminState: MutableState<Boolean>
     val showMainActivity: MutableLiveData<Boolean>
     val balance: MutableState<Double>
-    val qrCode: MutableState<Bitmap?>
 
     suspend fun loadData()
     suspend fun updateUser()
     suspend fun deleteUser()
     suspend fun getTotalBalance()
-    suspend fun shareQRCode()
 }
 
 @HiltViewModel
@@ -54,7 +49,6 @@ class ProfileViewModel @Inject constructor(
     override val toastChannel = Channel<UIText>()
     override val toast = toastChannel.receiveAsFlow()
     override val balance = mutableStateOf(0.0)
-    override val qrCode = mutableStateOf<Bitmap?>(null)
 
     init {
         viewModelScope.launch {
@@ -127,28 +121,6 @@ class ProfileViewModel @Inject constructor(
         balance.value = response.data.balance
     }
 
-    // reference: https://stackoverflow.com/questions/28232116/android-using-zxing-generate-qr-code
-    override suspend fun shareQRCode() {
-        val writer = QRCodeWriter()
-        val bitMatrix = writer.encode(
-            "https://github.com/morphclue",
-            BarcodeFormat.QR_CODE,
-            200,
-            200
-        )
-
-        val pixels = IntArray(bitMatrix.width * bitMatrix.height)
-        for (y in 0 until bitMatrix.height) {
-            for (x in 0 until bitMatrix.width) {
-                pixels[y * bitMatrix.width + x] = if (bitMatrix.get(x, y)) -0x1000000 else -0x1
-            }
-        }
-
-        val bitmap = Bitmap.createBitmap(bitMatrix.width, bitMatrix.height, Bitmap.Config.ARGB_8888)
-        bitmap.setPixels(pixels, 0, bitMatrix.width, 0, 0, bitMatrix.width, bitMatrix.height)
-        qrCode.value = bitmap
-    }
-
     private fun removeTags() {
         preference.removeTag(BuildConfig.USER_ID)
         preference.removeTag(BuildConfig.USER_PASSWORD)
@@ -167,7 +139,6 @@ class ProfileViewModelPreview : IProfileViewModel {
     override val balance = mutableStateOf(50.0)
     override val toastChannel = Channel<UIText>()
     override val toast = toastChannel.receiveAsFlow()
-    override val qrCode = mutableStateOf<Bitmap?>(null)
 
     override suspend fun loadData() {
         TODO("Not yet implemented")
@@ -182,10 +153,6 @@ class ProfileViewModelPreview : IProfileViewModel {
     }
 
     override suspend fun getTotalBalance() {
-        TODO("Not yet implemented")
-    }
-
-    override suspend fun shareQRCode() {
         TODO("Not yet implemented")
     }
 }
