+++ app/src/main/java/org/feature/fox/coffee_counter/ui/transaction/TransactionViewModel.kt
@@ -1,10 +1,13 @@
 package org.feature.fox.coffee_counter.ui.transaction
 
+import android.graphics.Bitmap
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.google.zxing.BarcodeFormat
+import com.google.zxing.qrcode.QRCodeWriter
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.receiveAsFlow
@@ -21,14 +24,15 @@ import org.feature.fox.coffee_counter.util.UIText
 import javax.inject.Inject
 
 interface ITransactionViewModel : IToast {
-
     val showMainActivity: MutableLiveData<Boolean>
     val toastMessage: MutableLiveData<String>
     val transactions: MutableList<TransactionResponse>
     val balance: MutableState<Double>
+    val qrCode: MutableState<Bitmap?>
 
     suspend fun refreshTransactions()
     suspend fun getTotalBalance()
+    suspend fun shareQRCode()
 }
 
 @HiltViewModel
@@ -36,13 +40,13 @@ class TransactionViewModel @Inject constructor(
     private val userRepository: UserRepository,
     private val preference: AppPreference,
 ) : ViewModel(), ITransactionViewModel {
-
     override val showMainActivity = MutableLiveData<Boolean>()
     override val toastMessage = MutableLiveData<String>()
     override val toastChannel = Channel<UIText>()
     override val toast = toastChannel.receiveAsFlow()
     override var transactions = mutableListOf<TransactionResponse>()
     override var balance = mutableStateOf(0.0)
+    override val qrCode = mutableStateOf<Bitmap?>(null)
 
     init {
         viewModelScope.launch {
@@ -84,6 +88,28 @@ class TransactionViewModel @Inject constructor(
         }
     }
 
+    // reference: https://stackoverflow.com/questions/28232116/android-using-zxing-generate-qr-code
+    override suspend fun shareQRCode() {
+        val writer = QRCodeWriter()
+        val bitMatrix = writer.encode(
+            "https://github.com/morphclue",
+            BarcodeFormat.QR_CODE,
+            200,
+            200
+        )
+
+        val pixels = IntArray(bitMatrix.width * bitMatrix.height)
+        for (y in 0 until bitMatrix.height) {
+            for (x in 0 until bitMatrix.width) {
+                pixels[y * bitMatrix.width + x] = if (bitMatrix.get(x, y)) -0x1000000 else -0x1
+            }
+        }
+
+        val bitmap = Bitmap.createBitmap(bitMatrix.width, bitMatrix.height, Bitmap.Config.ARGB_8888)
+        bitmap.setPixels(pixels, 0, bitMatrix.width, 0, 0, bitMatrix.width, bitMatrix.height)
+        qrCode.value = bitmap
+    }
+
     override suspend fun getTotalBalance() {
         val response = userRepository.getUserById(preference.getTag(BuildConfig.USER_ID))
 
@@ -105,6 +131,7 @@ class TransactionViewModelPreview : ITransactionViewModel {
     override val balance = mutableStateOf(13.0)
     override val toastChannel = Channel<UIText>()
     override val toast = toastChannel.receiveAsFlow()
+    override val qrCode = mutableStateOf<Bitmap?>(null)
 
     override suspend fun refreshTransactions() {}
 
@@ -112,4 +139,7 @@ class TransactionViewModelPreview : ITransactionViewModel {
         TODO("Not yet implemented")
     }
 
+    override suspend fun shareQRCode() {
+        TODO("Not yet implemented")
+    }
 }
